+++ CHANGELOG.md
@@ -6,6 +6,7 @@ in 1.x and 0.x versions.
 
 * 1.8.5
 
+  * Added flipped part detection
   * Added 'hit enter' behavior to validate forms in modal
   * Fixed some New Material modal issues